[Fix] Line Status Fix : S12P31S110-189

+++ BE/flip/src/main/java/com/ssafy/flip/domain/status/dto/request/LineSaveRequestDTO.java
@@ -1,7 +1,7 @@
 package com.ssafy.flip.domain.status.dto.request;
 
 public record LineSaveRequestDTO(
-        Long lineId,
+        Integer lineId,
         Float cycleTime,
         Boolean status
 ) {
